ml_ci_cd_python/requirements.txt
networkx
prometheus_client
pyyaml
scikit-learn
pandas
joblib
Flask
Flask-HTTPAuth
psutil
